Separating Integration Tests from Unit Tests (#9)
* test for byteToPtr

* splitting unit/integration testing

* splitting unit/integration testing
