+++ go.mod
@@ -1,11 +1,13 @@
 module github.com/randallmlough/pgxscan
 
-go 1.13
+go 1.15
 
 require (
 	github.com/jackc/pgx/v4 v4.1.2
-	github.com/lib/pq v1.2.0
-	github.com/pkg/errors v0.8.1
 	github.com/randallmlough/sqlmaper v0.0.0-20191117174101-7ad100a86097
 	github.com/stretchr/testify v1.4.0
+	golang.org/x/crypto v0.0.0-20200622213623-75b288015ac9 // indirect
+	golang.org/x/text v0.3.3 // indirect
+	golang.org/x/xerrors v0.0.0-20200804184101-5ec99f83aff1 // indirect
+	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
